Added ExitPlane and Audio Folder + Object Music
Added a Plane object that will check if collision is happening on it, which will send to the victory scene. Than I added a Music Folder with a Audio Object. There is a demo Music I input that can be active and use, or can be switch for a better suitable one.